Component added and UI changes
